+++ CHANGES.md
@@ -28,6 +28,7 @@ Changes in 6.0
 - Add MoodleXRosarioGet() function in Moodle/functions.php
 - Remove IsMoodleStudent(), IsMoodleUser(), IsMoodleCourse() & IsMoodleCoursePeriod() functions in Moodle/functions.php
 - Use MoodleXRosarioGet() in plugins/Moodle/* & Courses.php
+- HTML add arrow to indicate sub-profile in Profiles.php
 
 Changes in 5.9.4
 ----------------